feat: Update data structures to use Ustr for improved memory efficiency; refactor command handling and error enums for consistency
