+++ src/chart/models.rs
@@ -4,6 +4,7 @@ use chrono::{DateTime, Utc};
 use iso_currency::Currency;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+use ustr::Ustr;
 
 #[derive(Debug, Clone, Deserialize, Serialize, Copy, PartialEq, Eq, Hash)]
 pub enum ChartType {
@@ -60,18 +61,18 @@ impl PriceIterable for ChartHistoricalData {
     }
 }
 
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct ChartResponseData {
     #[serde(default)]
-    pub node: Option<String>,
+    pub node: Option<Ustr>,
     #[serde(rename(deserialize = "s"))]
     pub series: Vec<DataPoint>,
 }
 
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct StudyResponseData {
     #[serde(default)]
-    pub node: Option<String>,
+    pub node: Option<Ustr>,
     #[serde(rename(deserialize = "st"))]
     pub studies: Vec<DataPoint>,
     #[serde(rename(deserialize = "ns"))]
@@ -79,9 +80,9 @@ pub struct StudyResponseData {
 }
 
 // TODO: Implement graphic parser for indexes response
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct GraphicDataResponse {
-    pub d: String,
+    pub d: Ustr,
     pub indexes: Value,
 }
 
@@ -253,18 +254,18 @@ pub struct ChartDataChanges {
     pub zoffset: i64,
 }
 
-#[derive(Clone, PartialEq, Serialize, Hash, Debug, Default)]
+#[derive(Clone, PartialEq, Serialize, Hash, Debug, Default, Copy)]
 pub struct SeriesCompletedMessage {
     #[serde(default)]
-    pub id: String,
+    pub id: Ustr,
     #[serde(default)]
-    pub update_mode: String,
+    pub update_mode: Ustr,
 }
 
-#[derive(Clone, PartialEq, Serialize, Deserialize, Debug, Default, Builder)]
+#[derive(Clone, PartialEq, Serialize, Deserialize, Debug, Default, Builder, Copy)]
 pub struct MarketTicker {
-    pub symbol: String,
-    pub exchange: String,
+    pub symbol: Ustr,
+    pub exchange: Ustr,
     pub currency: Option<Currency>,
     pub country: Option<Currency>,
     pub market_type: Option<MarketType>,
@@ -273,8 +274,8 @@ pub struct MarketTicker {
 impl MarketTicker {
     pub fn new(symbol: &str, exchange: &str) -> Self {
         Self {
-            symbol: symbol.to_string(),
-            exchange: exchange.to_string(),
+            symbol: Ustr::from(symbol),
+            exchange: Ustr::from(exchange),
             currency: None,
             country: None,
             market_type: None,
@@ -297,8 +298,8 @@ impl MarketSymbol for MarketTicker {
 
     fn new<S: Into<String>>(symbol: S, exchange: S) -> Self {
         Self {
-            symbol: symbol.into(),
-            exchange: exchange.into(),
+            symbol: Ustr::from(&symbol.into()),
+            exchange: Ustr::from(&exchange.into()),
             currency: None,
             country: None,
             market_type: None,
@@ -328,29 +329,29 @@ impl From<SymbolInfo> for MarketTicker {
 #[serde(default)]
 pub struct SymbolInfo {
     #[serde(rename(deserialize = "pro_name"))]
-    pub id: String,
+    pub id: Ustr,
 
     #[serde(rename(deserialize = "original_name"))]
-    pub original_name: String,
+    pub original_name: Ustr,
 
-    pub name: String,
-    pub exchange: String,
-    pub description: String,
+    pub name: Ustr,
+    pub exchange: Ustr,
+    pub description: Ustr,
 
     #[serde(rename = "business_description")]
-    pub business_description: String,
+    pub business_description: Ustr,
 
     #[serde(rename = "listed_exchange")]
-    pub listed_exchange: String,
+    pub listed_exchange: Ustr,
 
     #[serde(rename = "provider_id")]
-    pub provider_id: String,
+    pub provider_id: Ustr,
 
     #[serde(rename = "base_currency")]
-    pub base_currency: String,
+    pub base_currency: Ustr,
 
     #[serde(rename = "base_currency_id")]
-    pub base_currency_id: String,
+    pub base_currency_id: Ustr,
 
     #[serde(rename = "total_revenue")]
     pub total_revenue: f64,
@@ -359,23 +360,23 @@ pub struct SymbolInfo {
     pub price_earnings_ttm: f64,
 
     #[serde(rename = "currency_id")]
-    pub currency_id: String,
+    pub currency_id: Ustr,
 
     #[serde(rename = "currency_code")]
-    pub currency_code: String,
+    pub currency_code: Ustr,
 
-    pub session_holidays: String,
+    pub session_holidays: Ustr,
 
     pub subsessions: Vec<Subsession>,
 
-    pub timezone: String,
+    pub timezone: Ustr,
 
     #[serde(rename(deserialize = "type"))]
-    pub market_type: String,
+    pub market_type: Ustr,
 
-    pub typespecs: Vec<String>,
+    pub typespecs: Vec<Ustr>,
 
-    pub aliases: Vec<String>,
+    pub aliases: Vec<Ustr>,
 
     pub total_shares_outstanding_calculated: f64,
 
@@ -383,11 +384,11 @@ pub struct SymbolInfo {
 
     pub earnings_release_date: i64,
 
-    pub base_name: Vec<String>,
+    pub base_name: Vec<Ustr>,
 
-    pub sector: String,
+    pub sector: Ustr,
 
-    pub current_session: String,
+    pub current_session: Ustr,
 
     pub founded: u16,
 
@@ -395,7 +396,7 @@ pub struct SymbolInfo {
 
     pub fractional: bool,
 
-    pub industry: String,
+    pub industry: Ustr,
 }
 
 impl MarketSymbol for SymbolInfo {
@@ -408,25 +409,25 @@ impl MarketSymbol for SymbolInfo {
     }
 
     fn id(&self) -> String {
-        self.id.clone()
+        self.id.to_string()
     }
 
     fn new<S: Into<String>>(symbol: S, exchange: S) -> Self {
         Self {
-            name: symbol.into(),
-            exchange: exchange.into(),
+            name: Ustr::from(&symbol.into()),
+            exchange: Ustr::from(&exchange.into()),
             ..Default::default()
         }
     }
 }
 
-#[derive(Clone, PartialEq, Serialize, Deserialize, Hash, Debug, Default)]
+#[derive(Clone, PartialEq, Serialize, Deserialize, Hash, Debug, Default, Copy)]
 #[serde(rename_all = "camelCase", default)]
 pub struct Subsession {
-    pub id: String,
-    pub description: String,
+    pub id: Ustr,
+    pub description: Ustr,
     pub private: bool,
-    pub session: String,
+    pub session: Ustr,
     #[serde(rename(deserialize = "session-display"))]
-    pub session_display: String,
+    pub session_display: Ustr,
 }
